tests/tasks/api/test_openrouter_tasks.py
import pytest
from unittest.mock import AsyncMock, patch
import asyncio
import os
from src.utils.settings import settings

from src.tasks.api.openrouter_tasks import complete_prompt

@pytest.fixture
def mock_openrouter_response():
    """Mock successful OpenRouter API response"""
    return {
        "id": "test-id",
        "model": "test-model",
        "created": 1234567890,
        "choices": [{
            "index": 0,
            "message": {
                "role": "assistant",
                "content": '{"test_prompt": "This is a test prompt."}'
            },
            "finish_reason": "stop"
        }],
        "usage": {
            "prompt_tokens": 10,
            "completion_tokens": 20,
            "total_tokens": 30
        }
    }

@pytest.fixture(autouse=True)
def mock_settings():
    """Mock settings to provide API key"""
    with patch('src.services.openrouter.settings') as mock_settings:
        # Use real API key from env if available, otherwise use test key
        api_key = os.getenv('OPENROUTER_API_KEY', 'test-key')
        print(f"Using API key: {api_key[:8]}...")  # Print first 8 chars for debugging
        mock_settings.OPENROUTER_API_KEY = api_key
        yield mock_settings

@pytest.mark.asyncio
async def test_complete_prompt_success(mock_openrouter_response, mock_settings):
    """Test successful API call with mocked response"""
    with patch('src.services.openrouter.OpenRouterClient') as MockClient:
        # Setup mock
        mock_client = AsyncMock()
        mock_client.complete.return_value = mock_openrouter_response
        MockClient.return_value = mock_client

        # Disable retries for testing
        with patch('src.tasks.api.openrouter_tasks.complete_prompt.retry_delay_seconds', 0):
            result = await complete_prompt(
                prompt="Test prompt",
                system_prompt="You are a test assistant",
                task_type="test"
            )

            # More flexible verification
            assert isinstance(result, dict)
            # Check that we got some kind of response
            assert len(result) > 0
            # Check that at least one value is a string
            assert any(isinstance(v, str) for v in result.values())

@pytest.mark.asyncio
async def test_complete_prompt_error():
    """Test error handling and retry behavior in the task"""
    with patch('src.services.openrouter.settings.OPENROUTER_API_KEY', 'test-key'), \
         patch('src.tasks.api.openrouter_tasks.OpenRouterClient') as MockClient:
        
        # Create mock instance
        mock_instance = AsyncMock()
        mock_instance.complete.side_effect = Exception("API Error")
        MockClient.return_value = mock_instance

        # Disable retries for testing
        with patch('src.tasks.api.openrouter_tasks.complete_prompt.retry_delay_seconds', 0):
            # Use async with for timeout
            async with asyncio.timeout(5):  # 5 second timeout
                with pytest.raises(Exception):
                    await complete_prompt("Test prompt")
                
                # Verify the mock was called 4 times (1 initial + 3 retries)
                assert mock_instance.complete.call_count == 4, "Should retry 3 times after initial failure"

@pytest.mark.integration
@pytest.mark.asyncio
async def test_complete_prompt_integration():
    """
    Integration test using actual OpenRouter API with free model
    Requires OPENROUTER_API_KEY environment variable
    """
    try:
        async with asyncio.timeout(10):  # Longer timeout for actual API call
            result = await complete_prompt(
                prompt="What is 1+1?",
                system_prompt="You are a helpful math tutor. Provide only the numerical answer.",
                task_type="test",  # Will use FREE_MODEL from constants
                temperature=0.1,
                max_tokens=10
            )
            
            assert result is not None
            assert any(char.isdigit() for char in str(result))
    except asyncio.TimeoutError:
        pytest.fail("Integration test timed out")
    except Exception as e:
        pytest.skip(f"Integration test skipped: {str(e)}") 